+++ Tests/Reqnroll.Formatters.Tests/Samples/Resources/NUnitRetry-retry-undefined/NUnitRetry_retry_undefined.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Reqnroll;
+
+namespace Reqnroll.Formatters.Tests.Samples.Resources.NUnitRetry_retry_undefined;
+
+[Binding]
+internal class NUnitRetry_retry_undefined
+{
+
+}