+++ Tests/Reqnroll.Formatters.Tests/Samples/Resources/minimal/minimal.ts
@@ -1,0 +1,5 @@
+import { Given } from '@cucumber/fake-cucumber'
+
+Given('I have {int} cukes in my belly', function (cukeCount: number) {
+  // no-op
+})